test(button): migrate basic
